6.code-interpreter/requirements.txt
openai
azure-ai-projects
azure-identity
python-dotenv
jsonref